+++ src/components/Settings/Header.js
@@ -1,16 +1,53 @@
-import { Image, Platform, Text, View } from "react-native";
-import { languageFlag, ScreensStyles } from "../../constants/styles";
+import { Platform, StyleSheet, View } from "react-native";
+import { ScreensStyles } from "../../constants/styles";
 import IconButton from "../UI/IconButton";
-import { useNavigation, useRouter } from "expo-router";
-import { useContext } from "react";
+import { useRouter } from "expo-router";
+import { useContext, useEffect, useState } from "react";
 import { StoryContext } from "../../context/stories-context";
 import { GlobalContext } from "../../context/global-context";
-import { findPropertyByKey, logos } from "../../constants/languages";
+import { LanguageOptionsNoLabel } from "../../constants/languages";
+import i18next from "i18next";
+import { PickerInput } from "../UI/PickerInput";
+
+export const HeaderLeft = () => {
+  const { globalConfig, setGlobalConfig } = useContext(GlobalContext);
+
+  const [inputAppLanguage, setInputAppLanguage] = useState(
+    i18next.language ? i18next.language : "en"
+  );
+
+  useEffect(() => {
+    // setIsFetching(true);
+    i18next.changeLanguage(inputAppLanguage);
+    setGlobalConfig({
+      ...globalConfig,
+      appLanguage: inputAppLanguage,
+    });
+    // setIsFetching(false);
+  }, [inputAppLanguage]);
+
+  useEffect(() => {
+    setInputAppLanguage(globalConfig.appLanguage);
+  }, [globalConfig.appLanguage]);
+
+  return (
+    <View style={styles.headerLeftView}>
+      <View style={{ width: Platform.OS == "web" ? "100%" : "25%" }}>
+        <PickerInput
+          onChangeText={(text) => {
+            setInputAppLanguage(text);
+          }}
+          zIndex={10000}
+          value={inputAppLanguage}
+          options={LanguageOptionsNoLabel}
+        />
+      </View>
+    </View>
+  );
+};
 
 export const HeaderRight = ({ tintColor }) => {
-  const { globalConfig, setGlobalConfig } = useContext(GlobalContext);
   const router = useRouter();
-  const { deleteStory } = useContext(StoryContext);
 
   function logoutHandler() {
     return (
@@ -42,3 +79,9 @@ export const HeaderRight = ({ tintColor }) => {
     </View>
   );
 };
+
+const styles = StyleSheet.create({
+  headerLeftView: {
+    marginLeft: 20,
+  },
+});